fixes for warnings with the sun studio compiler
git-svn-id: svn+ssh://svn.code.sf.net/p/qore/code/qore/trunk@3906 1b0b35fe-f740-4c1b-b404-a338b32e513b

+++ include/qore/intern/GlobalVariableList.h
@@ -53,7 +53,7 @@ public:
    DLLLOCAL Var *checkVar(const char *name, QoreParseTypeInfo *typeInfo, int *new_vars);
    DLLLOCAL Var *checkVar(const char *name, const QoreTypeInfo *typeInfo, int *new_vars);
    DLLLOCAL QoreListNode *getVarList() const;
-   DLLLOCAL void parseInit(int parse_options);
+   DLLLOCAL void parseInit(int64 parse_options);
    DLLLOCAL void parseCommit();
    DLLLOCAL void parseRollback();
 };
