+++ lib/QC_Datasource.cpp
@@ -302,7 +302,7 @@ static AbstractQoreNode *DS_getPort(QoreObject *self, ManagedDatasource *ds, con
 
 // Datasource::setTransactionLockTimeout(timeout $timeout_ms = 0) returns nothing
 static AbstractQoreNode *DS_setTransactionLockTimeout(QoreObject *self, ManagedDatasource *ds, const QoreListNode *params, ExceptionSink *xsink) {
-   ds->setTransactionLockTimeout(HARD_QORE_INT(params, 0));
+   ds->setTransactionLockTimeout((int)HARD_QORE_INT(params, 0));
    return 0;
 }
